AIFoundation/MLFoundation/Traditional/DenialPrediction/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Machine Learning
xgboost>=2.0.0
mlflow>=2.8.0
shap>=0.43.0
joblib>=1.3.0

# NLP & Transformers (optional for now)
# transformers>=4.35.0
# torch>=2.1.0
# tokenizers>=0.15.0

# Feature Store (optional for now)
# feast>=0.35.0

# Monitoring & Observability
prometheus-client>=0.19.0
evidently>=0.4.0

# Workflow Orchestration (optional for now)
# apache-airflow>=2.7.0
# temporalio>=1.4.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# API & Web
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development & Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# AWS/Cloud (optional)
# boto3>=1.34.0
# google-cloud-storage>=2.10.0

# Message Queue (optional for now)
# kafka-python>=2.0.0

# Security
cryptography>=41.0.0

# Date/Time
python-dateutil>=2.8.0

# JSON & Serialization
orjson>=3.9.0

# Logging
structlog>=23.2.0

# Configuration
pyyaml>=6.0.0

# Health Checks
healthcheck>=1.3.0

# Rate Limiting
slowapi>=0.1.0

# CORS
fastapi-cors>=0.0.6

# System Monitoring
psutil>=5.9.0 